Reorganize Contente Model code
