+++ demo/scripts/controls/contentModel/components/model/ContentModelImageView.tsx
@@ -11,7 +11,7 @@ import { MetadataView } from '../format/MetadataView';
 import { PaddingFormatRenderer } from '../format/formatPart/PaddingFormatRenderer';
 import { SegmentFormatView } from '../format/SegmentFormatView';
 import { SizeFormatRenderers } from '../format/formatPart/SizeFormatRenderers';
-import { updateImageMetadata } from 'roosterjs-content-model/lib/modelApi/metadata/updateImageMetadata';
+import { updateImageMetadata } from 'roosterjs-content-model';
 import { useProperty } from '../../hooks/useProperty';
 
 const styles = require('./ContentModelImageView.scss');
