+++ packages/roosterjs-content-model/lib/domUtils/constant.ts
@@ -1,0 +1,4 @@
+/**
+ * A placeholder text for color of hyperlink
+ */
+export const HyperLinkColorPlaceholder = '__hyperLinkColor';
