+++ packages/roosterjs-content-model/lib/domUtils/index.ts
@@ -1,0 +1,7 @@
+export { combineBorderValue, extractBorderValues } from './borderValues';
+export { updateImageMetadata } from './metadata/updateImageMetadata';
+export { updateListMetadata } from './metadata/updateListMetadata';
+export { updateMetadata } from './metadata/updateMetadata';
+export { updateTableCellMetadata } from './metadata/updateTableCellMetadata';
+export { updateTableMetadata } from './metadata/updateTableMetadata';
+export { HyperLinkColorPlaceholder } from './constant';
