+++ packages/roosterjs-content-model/lib/domUtils/metadata/updateImageMetadata.ts
