+++ packages/roosterjs-content-model/lib/domUtils/metadata/updateListMetadata.ts
