+++ packages/roosterjs-content-model/lib/domUtils/metadata/updateMetadata.ts
