+++ packages/roosterjs-content-model/lib/editor/index.ts
@@ -1,0 +1,3 @@
+export { default as ContentModelEditor } from './ContentModelEditor';
+export { default as isContentModelEditor } from './isContentModelEditor';
+export { default as ContentModelPlugin } from './ContentModelPlugin';
