+++ packages/roosterjs-content-model/lib/index.ts
@@ -1,184 +1,4 @@
-export { default as domToContentModel } from './publicApi/domToContentModel';
-export { default as contentModelToDom } from './publicApi/contentModelToDom';
-export { default as insertTable } from './publicApi/table/insertTable';
-export { default as formatTable } from './publicApi/table/formatTable';
-export { default as setTableCellShade } from './publicApi/table/setTableCellShade';
-export { default as editTable } from './publicApi/table/editTable';
-export { default as toggleBullet } from './publicApi/list/toggleBullet';
-export { default as toggleNumbering } from './publicApi/list/toggleNumbering';
-export { default as toggleBold } from './publicApi/segment/toggleBold';
-export { default as toggleItalic } from './publicApi/segment/toggleItalic';
-export { default as toggleUnderline } from './publicApi/segment/toggleUnderline';
-export { default as toggleStrikethrough } from './publicApi/segment/toggleStrikethrough';
-export { default as toggleSubscript } from './publicApi/segment/toggleSubscript';
-export { default as toggleSuperscript } from './publicApi/segment/toggleSuperscript';
-export { default as setBackgroundColor } from './publicApi/segment/setBackgroundColor';
-export { default as setFontName } from './publicApi/segment/setFontName';
-export { default as setFontSize } from './publicApi/segment/setFontSize';
-export { default as setTextColor } from './publicApi/segment/setTextColor';
-export { default as changeFontSize } from './publicApi/segment/changeFontSize';
-export { default as applySegmentFormat } from './publicApi/segment/applySegmentFormat';
-export { default as changeCapitalization } from './publicApi/segment/changeCapitalization';
-export { default as insertImage } from './publicApi/insert/insertImage';
-export { default as setListStyle } from './publicApi/list/setListStyle';
-export { default as setListStartNumber } from './publicApi/list/setListStartNumber';
-export { default as hasSelectionInBlock } from './publicApi/selection/hasSelectionInBlock';
-export { default as hasSelectionInSegment } from './publicApi/selection/hasSelectionInSegment';
-export { default as hasSelectionInBlockGroup } from './publicApi/selection/hasSelectionInBlockGroup';
-export { default as setIndentation } from './publicApi/block/setIndentation';
-export { default as setAlignment } from './publicApi/block/setAlignment';
-export { default as setDirection } from './publicApi/block/setDirection';
-export { default as setHeaderLevel } from './publicApi/block/setHeaderLevel';
-export { default as toggleBlockQuote } from './publicApi/block/toggleBlockQuote';
-export { default as setSpacing } from './publicApi/block/setSpacing';
-export { default as setImageBorder } from './publicApi/image/setImageBorder';
-export { default as setImageBoxShadow } from './publicApi/image/setImageBoxShadow';
-export { default as changeImage } from './publicApi/image/changeImage';
-export { default as getFormatState } from './publicApi/format/getFormatState';
-export { default as getSegmentFormat } from './publicApi/format/getSegmentFormat';
-export { default as clearFormat } from './publicApi/format/clearFormat';
-export { default as insertLink } from './publicApi/link/insertLink';
-export { default as removeLink } from './publicApi/link/removeLink';
-export { default as adjustLinkSelection } from './publicApi/link/adjustLinkSelection';
-export { default as setImageAltText } from './publicApi/image/setImageAltText';
-export { default as adjustImageSelection } from './publicApi/image/adjustImageSelection';
-
-export { default as ContentModelEditor } from './editor/ContentModelEditor';
-export { default as isContentModelEditor } from './editor/isContentModelEditor';
-
-export { default as ContentModelPlugin } from './publicPlugin/ContentModelPlugin';
-
-export { combineBorderValue, extractBorderValues } from './domUtils/borderValues';
-
-export { ContentModelBlockGroupType } from './publicTypes/enum/BlockGroupType';
-export { ContentModelBlockType } from './publicTypes/enum/BlockType';
-export { ContentModelSegmentType } from './publicTypes/enum/SegmentType';
-export { Selectable } from './publicTypes/selection/Selectable';
-
-export { ContentModelBlockBase } from './publicTypes/block/ContentModelBlockBase';
-export { ContentModelTable } from './publicTypes/block/ContentModelTable';
-export { ContentModelBlockGroupBase } from './publicTypes/group/ContentModelBlockGroupBase';
-export { ContentModelDocument } from './publicTypes/group/ContentModelDocument';
-export { ContentModelQuote } from './publicTypes/group/ContentModelQuote';
-export { ContentModelListItem } from './publicTypes/group/ContentModelListItem';
-export { ContentModelTableCell } from './publicTypes/group/ContentModelTableCell';
-export { ContentModelGeneralBlock } from './publicTypes/group/ContentModelGeneralBlock';
-export { ContentModelBlockGroup } from './publicTypes/group/ContentModelBlockGroup';
-export { ContentModelBlock } from './publicTypes/block/ContentModelBlock';
-export { ContentModelParagraph } from './publicTypes/block/ContentModelParagraph';
-export { ContentModelSegmentBase } from './publicTypes/segment/ContentModelSegmentBase';
-export { ContentModelSelectionMarker } from './publicTypes/segment/ContentModelSelectionMarker';
-export { ContentModelText } from './publicTypes/segment/ContentModelText';
-export { ContentModelBr } from './publicTypes/segment/ContentModelBr';
-export { ContentModelImage } from './publicTypes/segment/ContentModelImage';
-export { ContentModelGeneralSegment } from './publicTypes/segment/ContentModelGeneralSegment';
-export { ContentModelSegment } from './publicTypes/segment/ContentModelSegment';
-export { ContentModelEntity } from './publicTypes/entity/ContentModelEntity';
-export { ContentModelDivider } from './publicTypes/block/ContentModelDivider';
-
-export { ContentModelParagraphDecorator } from './publicTypes/decorator/ContentModelParagraphDecorator';
-export { ContentModelLink } from './publicTypes/decorator/ContentModelLink';
-
-export { FormatHandlerTypeMap, FormatKey } from './publicTypes/format/FormatHandlerTypeMap';
-export { ContentModelTableFormat } from './publicTypes/format/ContentModelTableFormat';
-export { ContentModelTableCellFormat } from './publicTypes/format/ContentModelTableCellFormat';
-export { ContentModelBlockFormat } from './publicTypes/format/ContentModelBlockFormat';
-export { ContentModelSegmentFormat } from './publicTypes/format/ContentModelSegmentFormat';
-export { ContentModelListItemLevelFormat } from './publicTypes/format/ContentModelListItemLevelFormat';
-export { ContentModelImageFormat } from './publicTypes/format/ContentModelImageFormat';
-export { ContentModelWithFormat } from './publicTypes/format/ContentModelWithFormat';
-export { ContentModelWithDataset } from './publicTypes/format/ContentModelWithDataset';
-export { ContentModelDividerFormat } from './publicTypes/format/ContentModelDividerFormat';
-
-export { VerticalAlignFormat } from './publicTypes/format/formatParts/VerticalAlignFormat';
-export { BackgroundColorFormat } from './publicTypes/format/formatParts/BackgroundColorFormat';
-export { BorderFormat } from './publicTypes/format/formatParts/BorderFormat';
-export { BorderBoxFormat } from './publicTypes/format/formatParts/BorderBoxFormat';
-export { IdFormat } from './publicTypes/format/formatParts/IdFormat';
-export { SizeFormat } from './publicTypes/format/formatParts/SizeFormat';
-export { SpacingFormat } from './publicTypes/format/formatParts/SpacingFormat';
-export { DirectionFormat } from './publicTypes/format/formatParts/DirectionFormat';
-export { TextColorFormat } from './publicTypes/format/formatParts/TextColorFormat';
-export { FontSizeFormat } from './publicTypes/format/formatParts/FontSizeFormat';
-export { FontFamilyFormat } from './publicTypes/format/formatParts/FontFamilyFormat';
-export { BoldFormat } from './publicTypes/format/formatParts/BoldFormat';
-export { ItalicFormat } from './publicTypes/format/formatParts/ItalicFormat';
-export { UnderlineFormat } from './publicTypes/format/formatParts/UnderlineFormat';
-export { StrikeFormat } from './publicTypes/format/formatParts/StrikeFormat';
-export { SuperOrSubScriptFormat } from './publicTypes/format/formatParts/SuperOrSubScriptFormat';
-export { TableMetadataFormat } from './publicTypes/format/formatParts/TableMetadataFormat';
-export { ContentModelFormatBase } from './publicTypes/format/ContentModelFormatBase';
-export { MarginFormat } from './publicTypes/format/formatParts/MarginFormat';
-export { PaddingFormat } from './publicTypes/format/formatParts/PaddingFormat';
-export { DisplayFormat } from './publicTypes/format/formatParts/DisplayFormat';
-export { LineHeightFormat } from './publicTypes/format/formatParts/LineHeightFormat';
-export { LinkFormat } from './publicTypes/format/formatParts/LinkFormat';
-export { ListTypeFormat } from './publicTypes/format/formatParts/ListTypeFormat';
-export { ListThreadFormat } from './publicTypes/format/formatParts/ListThreadFormat';
-export { ListMetadataFormat } from './publicTypes/format/formatParts/ListMetadataFormat';
-export {
-    ImageResizeMetadataFormat,
-    ImageCropMetadataFormat,
-    ImageMetadataFormat,
-    ImageRotateMetadataFormat,
-} from './publicTypes/format/formatParts/ImageMetadataFormat';
-export { DatasetFormat } from './publicTypes/format/formatParts/DatasetFormat';
-export { WhiteSpaceFormat } from './publicTypes/format/formatParts/WhiteSpaceFormat';
-export { WordBreakFormat } from './publicTypes/format/formatParts/WordBreakFormat';
-export { ZoomScaleFormat } from './publicTypes/format/formatParts/ZoomScaleFormat';
-
-export { ContentModelFormatMap } from './publicTypes/format/ContentModelFormatMap';
-
-export { EditorContext } from './publicTypes/context/EditorContext';
-export {
-    DomToModelListFormat,
-    DomToModelFormatContext,
-} from './publicTypes/context/DomToModelFormatContext';
-export {
-    DomToModelRegularSelection,
-    DomToModelTableSelection,
-    DomToModelImageSelection,
-    DomToModelSelectionContext,
-} from './publicTypes/context/DomToModelSelectionContext';
-export {
-    DomToModelSettings,
-    DefaultStyleMap,
-    ElementProcessorMap,
-    FormatParser,
-    FormatParsers,
-    FormatParsersPerCategory,
-} from './publicTypes/context/DomToModelSettings';
-export { DomToModelContext } from './publicTypes/context/DomToModelContext';
-export { ModelToDomContext } from './publicTypes/context/ModelToDomContext';
-export {
-    ModelToDomListStackItem,
-    ModelToDomListContext,
-    ModelToDomFormatContext,
-} from './publicTypes/context/ModelToDomFormatContext';
-export {
-    ModelToDomBlockAndSegmentNode,
-    ModelToDomRegularSelection,
-    ModelToDomTableSelection,
-    ModelToDomImageSelection,
-    ModelToDomSelectionContext,
-} from './publicTypes/context/ModelToDomSelectionContext';
-export {
-    ModelToDomSettings,
-    FormatApplier,
-    FormatAppliers,
-    FormatAppliersPerCategory,
-    ContentModelHandlerMap,
-    ContentModelHandlerTypeMap,
-    DefaultImplicitFormatMap,
-} from './publicTypes/context/ModelToDomSettings';
-export { ModelToDomEntityContext } from './publicTypes/context/ModelToDomEntityContext';
-export { ElementProcessor } from './publicTypes/context/ElementProcessor';
-export { ContentModelHandler } from './publicTypes/context/ContentModelHandler';
-
-export { Border } from './publicTypes/interface/Border';
-
-export {
-    IContentModelEditor,
-    DomToModelOption,
-    ModelToDomOption,
-} from './publicTypes/IContentModelEditor';
+export * from './publicTypes/index';
+export * from './editor/index';
+export * from './publicApi/index';
+export * from './domUtils/index';
