+++ packages/roosterjs-content-model/lib/modelApi/common/insertContent.ts
@@ -1,4 +1,4 @@
-import domToContentModel from '../../publicApi/domToContentModel';
+import domToContentModel from '../../domToModel/domToContentModel';
 import { ContentModelDocument } from '../../publicTypes/group/ContentModelDocument';
 import { mergeModel } from '../../modelApi/common/mergeModel';
 import { safeInstanceOf, wrap } from 'roosterjs-editor-dom';
