+++ packages/roosterjs-content-model/lib/publicApi/image/insertImage.ts
