+++ packages/roosterjs-content-model/lib/publicApi/index.ts
@@ -1,0 +1,42 @@
+export { default as insertTable } from './table/insertTable';
+export { default as formatTable } from './table/formatTable';
+export { default as setTableCellShade } from './table/setTableCellShade';
+export { default as editTable } from './table/editTable';
+export { default as toggleBullet } from './list/toggleBullet';
+export { default as toggleNumbering } from './list/toggleNumbering';
+export { default as toggleBold } from './segment/toggleBold';
+export { default as toggleItalic } from './segment/toggleItalic';
+export { default as toggleUnderline } from './segment/toggleUnderline';
+export { default as toggleStrikethrough } from './segment/toggleStrikethrough';
+export { default as toggleSubscript } from './segment/toggleSubscript';
+export { default as toggleSuperscript } from './segment/toggleSuperscript';
+export { default as setBackgroundColor } from './segment/setBackgroundColor';
+export { default as setFontName } from './segment/setFontName';
+export { default as setFontSize } from './segment/setFontSize';
+export { default as setTextColor } from './segment/setTextColor';
+export { default as changeFontSize } from './segment/changeFontSize';
+export { default as applySegmentFormat } from './segment/applySegmentFormat';
+export { default as changeCapitalization } from './segment/changeCapitalization';
+export { default as insertImage } from './image/insertImage';
+export { default as setListStyle } from './list/setListStyle';
+export { default as setListStartNumber } from './list/setListStartNumber';
+export { default as hasSelectionInBlock } from './selection/hasSelectionInBlock';
+export { default as hasSelectionInSegment } from './selection/hasSelectionInSegment';
+export { default as hasSelectionInBlockGroup } from './selection/hasSelectionInBlockGroup';
+export { default as setIndentation } from './block/setIndentation';
+export { default as setAlignment } from './block/setAlignment';
+export { default as setDirection } from './block/setDirection';
+export { default as setHeaderLevel } from './block/setHeaderLevel';
+export { default as toggleBlockQuote } from './block/toggleBlockQuote';
+export { default as setSpacing } from './block/setSpacing';
+export { default as setImageBorder } from './image/setImageBorder';
+export { default as setImageBoxShadow } from './image/setImageBoxShadow';
+export { default as changeImage } from './image/changeImage';
+export { default as getFormatState } from './format/getFormatState';
+export { default as getSegmentFormat } from './format/getSegmentFormat';
+export { default as clearFormat } from './format/clearFormat';
+export { default as insertLink } from './link/insertLink';
+export { default as removeLink } from './link/removeLink';
+export { default as adjustLinkSelection } from './link/adjustLinkSelection';
+export { default as setImageAltText } from './image/setImageAltText';
+export { default as adjustImageSelection } from './image/adjustImageSelection';
