+++ packages/roosterjs-content-model/test/formatHandlers/segment/textColorFormatHandlerTest.ts
@@ -1,7 +1,7 @@
 import { createDomToModelContext } from '../../../lib/domToModel/context/createDomToModelContext';
 import { createModelToDomContext } from '../../../lib/modelToDom/context/createModelToDomContext';
 import { DomToModelContext } from '../../../lib/publicTypes/context/DomToModelContext';
-import { HyperLinkColorPlaceholder } from '../../../lib/formatHandlers/utils/defaultStyles';
+import { HyperLinkColorPlaceholder } from '../../../lib/domUtils/constant';
 import { ModelToDomContext } from '../../../lib/publicTypes/context/ModelToDomContext';
 import { TextColorFormat } from '../../../lib/publicTypes/format/formatParts/TextColorFormat';
 import { textColorFormatHandler } from '../../../lib/formatHandlers/segment/textColorFormatHandler';
