+++ packages/roosterjs-content-model/test/publicApi/segment/changeFontSizeTest.ts
@@ -1,7 +1,7 @@
 import * as getComputedStyles from 'roosterjs-editor-dom/lib/utils/getComputedStyles';
 import * as pendingFormat from '../../../lib/publicApi/format/pendingFormat';
 import changeFontSize from '../../../lib/publicApi/segment/changeFontSize';
-import domToContentModel from '../../../lib/publicApi/domToContentModel';
+import domToContentModel from '../../../lib/domToModel/domToContentModel';
 import { ContentModelDocument } from '../../../lib/publicTypes/group/ContentModelDocument';
 import { createRange } from 'roosterjs-editor-dom';
 import { IContentModelEditor } from '../../../lib/publicTypes/IContentModelEditor';
